Refactor: Sections 일급컬렉션 처리
